revision previa a prubeas y aplicacion d e validaciones restantes al progress Log

+++ src/main/java/acme/features/client/dashboard/ClientDashboardShowService.java
@@ -38,9 +38,9 @@ public class ClientDashboardShowService extends AbstractService<Client, ClientDa
 		int clientId;
 
 		clientId = super.getRequest().getPrincipal().getActiveRoleId();
-		float percentaje25 = 25.0f;
-		float percentaje50 = 50.0f;
-		float percentaje75 = 75.0f;
+		double percentaje25 = 25.0;
+		double percentaje50 = 50.0;
+		double percentaje75 = 75.0;
 
 		totalLogsWithCompletenessBelow25 = this.repository.logsBelowCompletenessValue(clientId, percentaje25);
 		totalLogsWithCompletenessBetween25And50 = this.repository.logsBetweenCompletenessValuesForClient(clientId, percentaje25, percentaje50);
